# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Personal banking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send email verification OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, message]
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  devCode:
                    type: string
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify email verification OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, message]
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        "400":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /accounts/{id}/statement:
    get:
      operationId: getAccountStatement
      tags: [accounts]
      summary: Get account statement (transactions in date range)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    RegisterInput:
      type: object
      required: [firstName, lastName, email, password, phone]
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        phone:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    SendOtpInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    VerifyOtpInput:
      type: object
      required: [email, code]
      properties:
        email:
          type: string
          format: email
        code:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
    UserProfile:
      type: object
      required: [id, firstName, lastName, email, phone, isEmailVerified, createdAt]
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        isEmailVerified:
          type: boolean
        avatarUrl:
          type: ["string", "null"]
        bvn:
          type: ["string", "null"]
        nin:
          type: ["string", "null"]
        kycStatus:
          type: string
          enum: [pending, verified, failed]
        createdAt:
          type: string
          format: date-time
...
    Account:
      type: object
      required: [id, userId, accountNumber, type, balance, currency, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        accountNumber:
          type: string
        type:
          type: string
          enum: [current, savings, wallet]
        balance:
          type: number
        currency:
          type: string
          default: NGN
        status:
          type: string
          enum: [active, frozen, closed]
        bankName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    Transaction:
      type: object
      required: [id, accountId, type, amount, currency, description, status, reference, createdAt]
      properties:
        id:
          type: string
        accountId:
          type: string
        type:
          type: string
          enum: [credit, debit]
        amount:
          type: number
        currency:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, completed, failed, reversed]
        reference:
          type: string
        category:
          type: ["string", "null"]
        recipientName:
          type: ["string", "null"]
        recipientBank:
          type: ["string", "null"]
        recipientAccount:
          type: ["string", "null"]
        senderName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...